src/backends/requirements.txt
fastapi
undetected-playwright-patch
openai
Pillow
mcp
uvicorn
undetected_playwright